repomind-backend/requirements.txt
fastapi
uvicorn
pydantic-settings
groq
asyncpg
psycopg2-binary
sqlalchemy
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface 
langchain-postgres
langchain-core
langgraph
gitpython
bcrypt
python-jose